fastapi/requirements.txt
fastapi
uvicorn[standard]
watchdog
sqlalchemy
mysql-connector-python
pydantic
alembic
python-dotenv
Pillow